Better solution to class change on refresh
